**Remove unnecessary files and directories from the Setting module**
This commit message is for a situation where you need to remove some files and directories from the Setting module in your Laravel project. The reason for this might be that these files are no longer needed or they were created by mistake.

Here's what I did:

* Removed all the unnecessary files and directories from the `Setting` module.
* This includes the `tests`, `vite.config.js`, `webpack.mix.js`, `tailwind.config.js`, `testbench.yaml`, and some other files that are not needed anymore.
* The commit message is clear about what was removed, so it's easy to track changes in the project.

This change will help keep the project organized and easier to maintain.